+++ meta.go
@@ -15,7 +15,7 @@ import (
 )
 
 var (
-	version = "3.2.0"
+	version = "3.3.0"
 	commit  = "unknown"
 	builtAt = "unknown"
 )